First Draft Domains
These domains are based on the domains from the following publications:

- 1-5-2-3-4-full: Kabanza et al. Controlling the hypothesis space in probabilistic plan recognition. IJCAI. 2013.
A simulated domain based on and/or trees. The values in the name represent the characretirsics of the domain - 1 plan execution, 5 possible goals, or-branching of 2, and-branching of 3, depth of 4, fully ordered. This domain has no parameters.

- Monroe: Blaylock. Towards tractable agent-based dialogue. Diss. University of Rochester, 2005.
A domain based on plan executions of disaster response teams. Parameterized domain, partial order.

- Soccer: Avrahami-Zilberbrand, and Kaminka. Incorporating observer biases in keyhole plan recognition (efficiently!). AAAI. Vol. 7. 2007.
A domain based on the robo-cup planning competition. This is a very simple domain,  no parameters and fully ordered.

-TinkerPlots: Oriel Uzan, Reuth Dekel, Or Seri, Ya'akov Gal.  Plan Recognition for Exploratory Learning